+++ XCDYouTubeKit Demo/iOS Demo/DemoFullScreenViewController.h
@@ -1,8 +1,10 @@
 //
-//  Copyright (c) 2013-2014 Cédric Luthi. All rights reserved.
+//  Copyright (c) 2013-2015 Cédric Luthi. All rights reserved.
 //
 
-@interface DemoFullScreenViewController : UIViewController <UITextFieldDelegate>
+#import "VideoPickerController.h"
+
+@interface DemoFullScreenViewController : UIViewController <UITextFieldDelegate, VideoPickerControllerDelegate>
 
 @property (nonatomic, weak) IBOutlet UITextField *videoIdentifierTextField;
 @property (nonatomic, weak) IBOutlet UISwitch *lowQualitySwitch;