+++ XCDYouTubeKit/XCDYouTubeVideoPlayerViewController.m
@@ -1,5 +1,5 @@
 //
-//  Copyright (c) 2013-2014 Cédric Luthi. All rights reserved.
+//  Copyright (c) 2013-2015 Cédric Luthi. All rights reserved.
 //
 
 #import "XCDYouTubeVideoPlayerViewController.h"
@@ -160,7 +160,7 @@ NSString *const XCDYouTubeVideoUserInfoKey = @"Video";
 
 - (void) presentInView:(UIView *)view
 {
-	static void *XCDYouTubeVideoPlayerViewControllerKey = &XCDYouTubeVideoPlayerViewControllerKey;
+	static const void * const XCDYouTubeVideoPlayerViewControllerKey = &XCDYouTubeVideoPlayerViewControllerKey;
 	
 	self.embedded = YES;